orderscafe/myproject/requirements.txt
asgiref==3.8.1
black==25.1.0
click==8.1.8
Django==5.1.7
djangorestframework==3.15.2
flake8==7.1.2
isort==6.0.1
mccabe==0.7.0
mypy==1.15.0
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.7
pycodestyle==2.12.1
pyflakes==3.2.0
python-dotenv==1.1.0
sqlparse==0.5.3
typing_extensions==4.13.0
